Add files via upload
The pass_log is a database created to save the pass words of the users and admins of  the Football Sports League Management System(hence forth referred to as fbslms) database. The login page checks from the user the username and password passing an error message when they are not in the database. Further there is an option to choose usage as a fan or as an admin and the error message when no access present. Also the first iterations are purely built from a functionability standpoint.